third_party/blink/web_tests/fast/html/imports/import-element-removed-flag-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
This tests 'element removed flag' behavior defined in https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/imports/index.html#dfn-element-removed-flag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS staticImport.import is non-null.
PASS staticImport.import is null
PASS staticImport.import is non-null.
PASS dynamicImport.import is non-null.
PASS dynamicImport.import is null
PASS dynamicImport.import is non-null.
PASS dynamicImportEager.import is null
PASS dynamicImportEager.import is non-null.
PASS successfullyParsed is true

TEST COMPLETE
